Working Transmission system!!
Used files from lab3-ex3 as an example to make a working transmission system from.  Can choose your move and then send it to the other play, when you receive a valid move then your blue led turns on. When both players have the blue led on then they can press the other button on the board, at which point the led turns off and the other players move displays on your screen.